requirements.txt
apiclient==1.0.3
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
facebook-sdk==2.0.0
Flask==0.12.2
foursquare==1!2016.9.12
future==0.16.0
google-api-python-client==1.6.6
httplib2==0.11.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
langdetect==1.0.7
Lemmagen==1.2.0
lxml==4.2.1
MarkupSafe==1.0
mysql-connector-python==8.0.6
nltk==3.2.5
numpy==1.14.2
oauth2==1.9.0.post1
oauth2client==4.1.2
oauthlib==2.0.7
pandas==0.22.0
progressbar==2.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pygame==1.9.3
PySocks==1.6.8
pytagcloud==0.3.5
python-dateutil==2.7.2
PyTumblr==0.0.7
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
simplejson==3.13.2
six==1.11.0
tweepy==3.6.0
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.14.1
